requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.9
idna==3.11
psycopg2-binary==2.9.11
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.2
